basic: fix prog_* uint8_t line-counter overflow (hung for >255 lines)
The prog_* line-table helpers scanned the program with a uint8_t loop
counter: for (uint8_t i; i < TIKU_BASIC_PROGRAM_LINES; i++). A uint8_t
caps at 255, so once 0411cac raised PROGRAM_LINES per tier (1024 on
Apollo4 Lite / RP2350, 2048 on Apollo510, 256 on msp430-large) the
comparison stayed permanently true and the scan looped forever. It hung
when the first numbered line was stored -- prog_store never returned, so
the REPL prompt never came back and the next-typed RUN looked like the
culprit. Only the host build (PROGRAM_LINES=50) was unaffected.

Widen the five scan counters to uint16_t (prog_clear, prog_store,
prog_next_index, prog_find_exact in tiku_basic_program.inl, and
prog_find_label in tiku_basic_stmt.inl; line counts and numbers are
already uint16_t) and add a _Static_assert so a future PROGRAM_LINES
bump past 65535 cannot silently reintroduce it.

Diagnosed by JLink attach-halt on apollo4l: IPSR=NoException with the PC
pinned in prog_store == a busy loop, not a fault. Verified on apollo4l
hardware: storing lines, RUN (42), GOTO-to-label and LIST all work and
the board stays responsive.

Also correct file-store-handoff.md, which had mislabeled this as a
fault / RUN-only / Apollo D-cache "Bug #1" blocker.

+++ kernel/shell/basic/tiku_basic_stmt.inl
@@ -394,7 +394,7 @@ exec_input(const char **p)
 static int
 prog_find_label(const char *name, size_t name_len)
 {
-    uint8_t i;
+    uint16_t i;
     size_t  k;
     for (i = 0; i < TIKU_BASIC_PROGRAM_LINES; i++) {
         const char *t;
